+++ Boxes/EsdsBox.cs
@@ -8,7 +8,7 @@ namespace AAXClean.Boxes
     //https://developer.apple.com/library/archive/documentation/QuickTime/QTFF/QTFFChap3/qtff3.html#//apple_ref/doc/uid/TP40000939-CH205-124774
     internal class EsdsBox : FullBox
     {
-        public override uint RenderSize => base.RenderSize + ES_Descriptor.RenderSize;
+        public override long RenderSize => base.RenderSize + ES_Descriptor.RenderSize;
 
         public ES_Descriptor ES_Descriptor { get; }
         internal EsdsBox(Stream file, BoxHeader header, Box parent) : base(file, header, parent)
