+++ Boxes/FullBox.cs
@@ -5,7 +5,7 @@ namespace AAXClean.Boxes
 {
     internal abstract class FullBox : Box, IFullBox
     {
-        public override uint RenderSize => base.RenderSize + 4;
+        public override long RenderSize => base.RenderSize + 4;
         public byte Version => VersionFlags[0];
         public int Flags => VersionFlags[1] << 16 | VersionFlags[2] << 8 | VersionFlags[3];
 
@@ -26,6 +26,6 @@ namespace AAXClean.Boxes
 
     internal interface IFullBox
     {
-        uint RenderSize { get; }
+        long RenderSize { get; }
     }
 }
